Fix unit test compiling on Linux Review URL: http://codereview.chromium.org/295034
git-svn-id: svn://svn.chromium.org/chrome/trunk/src@29704 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 00bb04380cd76943ec66366e315122f9f7670b98

+++ chrome/browser/sync/notifier/communicator/single_login_attempt.cc
@@ -530,7 +530,7 @@ void SingleLoginAttempt::HandleConnectionError(
         // Yep, its a "stream:error" with "see-other-host" text, let's parse
         // out the server:port, and then reconnect with that.
         const std::string& redirect = text->BodyText();
-        unsigned int colon = redirect.find(":");
+        size_t colon = redirect.find(":");
         int redirect_port = kDefaultXmppPort;
         std::string redirect_server;
         if (colon == std::string::npos) {
